part3/requirements.txt
# Core Flask and extensions
Flask==2.3.3
Flask-RESTx==1.3.0
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1

# Database and ORM
SQLAlchemy==2.0.43

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.2.1
pytest-flask==1.3.0

# Development tools
flask-cors==4.0.0

# Additional utilities that might be needed
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
PyJWT==2.8.0
bcrypt==4.0.1
greenlet==2.0.2
typing-extensions==4.8.0

# Optional: For better development experience
python-dateutil==2.8.2